Check for missing or duplicated anchors, rework v2
