+++ yamllint/conf/default.yaml
@@ -6,7 +6,7 @@ yaml-files:
   - '.yamllint'
 
 rules:
-  anchor-duplicates: disable
+  anchors: enable
   braces: enable
   brackets: enable
   colons: enable
@@ -33,4 +33,3 @@ rules:
   trailing-spaces: enable
   truthy:
     level: warning
-  unknown-aliases: enable
